+++ RELEASE_NOTES.md
@@ -5,6 +5,11 @@ A new header is inserted each time a *tag* is created.
 
 ## main branch
 
+## v1.28.3
+
+- backend: add support for GGP platform
+- engine: primitives with `CullingMode::FRONT_AND_BACK` are now skipped.
+
 ## v1.28.2
 
 - gltfio: support EXT_meshopt_compression